owtimeslice: Hotfix repainting
On click of button, buttons do not repaint without a call to repaint(),
even with an explicit update() call. Same goes for changing the page of
a QDateTimeEdit by clicking one of the arrows in the popup calendar.
It does this on Qt5.12 through Qt5.14 on Mac.

This widget is a bit weird, the underlying cause should be found.